rust: arc: make examples compile
This is part of the effort to minimize the differences of the `rust`
branch with respect to mainline in order to eventually drop it.

This commit increases the difference, but it is required to make
the examples compile (which we do not do yet in mainline). It will
land in mainline eventually.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>